+++ packages/uif/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @l2beat/uif
 
+## 0.4.0
+
+### Minor Changes
+
+- Add getSafeHeight and setSafeHeight to MultiIndexer
+
+## 0.3.1
+
+### Patch Changes
+
+- Add ability to provide configurations from different source to MultiIndexer
+
 ## 0.3.0
 
 ### Minor Changes